+++ src/dotNet/FormPet.cs
@@ -673,17 +673,19 @@ namespace DesktopPet
                                 // window right border!
                                 if (PositionX + x <= rct.Right
                                     && PositionX + x > rct.Right + x // A little bit of wiggle-room so the sheep doesn't phase through the window
+                                    && PositionY + y < rct.Bottom // Ignore if below window. REMINDER: Y is inverted
+                                    && PositionY + y + Height > rct.Top // Ignore if above window
                                     && !CheckTopWindow(false, window.Key) // Ignore windows that aren't visible on top
                                     && rct.Right < ScreenArea.Right - Width // Ignore collision if the sheep doesn't have enough space to make it past the right screen border (ex: if an application is covering the entire screen)
                                     )
                                 {
-                                    //TODO: remove/simplify this debug
                                     StartUp.AddDebugInfo(StartUp.DEBUG_TYPE.warning, window.Value + " right-side collision!");
-                                    Console.WriteLine("'" + window.Value + "' Radius Right: " + rct.Right + ", Left: " + (rct.Right + x) + " | Sheep Left: " + (PositionX + x));
+
+                                    Console.WriteLine("\n" + CurrentAnimation.Name);
 
                                     //TODO: add "only" functionality for WINDOWSIDE and/or PETSIDE value
                                     // Set border animation
-                                    int iBorderAnimation = Animations.SetNextBorderAnimation(CurrentAnimation.ID, TNextAnimation.TOnly.NONE);
+                                    int iBorderAnimation = Animations.SetNextBorderAnimation(CurrentAnimation.ID, TNextAnimation.TOnly.VERTICAL);
                                     // If a "AAAAAAAAAAAAAAAA" or "NONE" border animation exists, play it. Otherwise, pet ignores this collision and continues as normal.
                                     if (iBorderAnimation >= 0)
                                     {
@@ -741,6 +743,47 @@ namespace DesktopPet
                         else
                         {
                             bLeavingScreen = true;
+                        }
+                    }
+                    else    // run window border detection setup and then check left border of all valid windows!
+                    {
+                        // Retrieve valid windows
+                        Dictionary<IntPtr, string> windows = EnumerateWindows();
+
+                        // For each valid window found:
+                        foreach (KeyValuePair<IntPtr, string> window in windows)
+                        {
+                            // Get size and position of window
+                            if (NativeMethods.GetWindowRect(new HandleRef(this, window.Key), out NativeMethods.RECT rct))
+                            {
+                                // window left border!
+                                if (PositionX + x + Width >= rct.Left
+                                    && PositionX + x + Width < rct.Left + x // A little bit of wiggle-room so the sheep doesn't phase through the window
+                                    && PositionY + y < rct.Bottom // Ignore if below window. REMINDER: Y is inverted
+                                    && PositionY + y + Height > rct.Top // Ignore if above window
+                                    && !CheckTopWindow(false, window.Key) // Ignore windows that aren't visible on top
+                                    && rct.Left > ScreenArea.Left + Width // Ignore collision if the sheep doesn't have enough space to make it past the left screen border (ex: if an application is covering the entire screen)
+                                    )
+                                {
+                                    StartUp.AddDebugInfo(StartUp.DEBUG_TYPE.warning, window.Value + " left-side collision!");
+
+                                    Console.WriteLine("\n" + CurrentAnimation.Name);
+
+                                    //TODO: add "only" functionality for WINDOWSIDE and/or PETSIDE value
+                                    // Set border animation
+                                    int iBorderAnimation = Animations.SetNextBorderAnimation(CurrentAnimation.ID, TNextAnimation.TOnly.VERTICAL);
+                                    // If a "AAAAAAAAAAAAAAAA" or "NONE" border animation exists, play it. Otherwise, pet ignores this collision and continues as normal.
+                                    if (iBorderAnimation >= 0)
+                                    {
+                                        PositionX = rct.Left - Width;
+                                        x = 0;
+                                        SetNewAnimation(iBorderAnimation);
+                                        bNewAnimation = true;
+                                    }
+
+                                    break;
+                                }
+                            }
                         }
                     }
                 }